feat: add KVMetadata via shared BaseMetadata refactor (#153)
Extract the provider-agnostic metadata logic from MLYMetadata into a new
BaseMetadata class; MLYMetadata and KVMetadata are now thin subclasses that
only supply input loading + column normalization. KVMetadata maps KartaView
kv_pids.csv columns (shotDate->captured_at ms, heading->compass_angle,
sequenceId->sequence_id, userId->creator_id, orgCode->organization_id) to the
base schema, so KartaView datasets get the same indicators as Mapillary.

Also fix compute_metadata reloading the raw CSV directly (which discarded
normalization); it now reloads through _normalize_columns. MLYMetadata behavior
is unchanged (its normalization is a no-op) and test_mly_metadata.py stays green.

Adds test_kv_metadata.py + a kv_pids.csv fixture. Bumps minor version to 1.5.0
(new KVMetadata/BaseMetadata public API).

Closes #153
