+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "zensvi"
-version = "1.4.9"
+version = "1.5.0"
 description = "This package handles downloading, cleaning, analyzing street view imagery in a one-stop and zen manner."
 authors = ["koito19960406"]
 license = "MIT"
