+++ src/zensvi/metadata/__init__.py
@@ -1,3 +1,5 @@
+from .base_metadata import BaseMetadata
+from .kv_metadata import KVMetadata
 from .mly_metadata import MLYMetadata
 
-__all__ = ["MLYMetadata"]
+__all__ = ["BaseMetadata", "KVMetadata", "MLYMetadata"]
