+++ tests/data/input/metadata/kv_pids.csv
@@ -1,0 +1,7 @@
+id,lat,lon,shotDate,heading,sequenceId,userId,orgCode,is_pano
+1871849321,1.313448531810593,103.79303455352783,2023-03-06 02:29:12.000,178.57,8326665,44,CMNT,True
+1871849322,1.3128639638087236,103.7926858663559,2023-03-06 02:29:14.000,180.10,8326665,44,CMNT,True
+1871849323,1.3140116476463817,103.79172563552856,2022-07-11 09:15:03.000,90.00,8326666,44,CMNT,False
+1871849324,1.313980,103.791800,2022-07-11 09:15:07.000,92.50,8326666,51,ACME,False
+1871849325,1.313500,103.792900,2021-12-20 23:45:00.000,270.00,8326667,51,ACME,True
+1871849326,1.313100,103.792700,2021-12-20 23:45:05.000,268.20,8326667,51,,True
